+++ sorellaFront/src/app/app-routing.module.ts
@@ -8,11 +8,11 @@ const routes: Routes = [
   },
   {
     path: 'login',
-    loadChildren: () => import('./login/login.module').then( m => m.LoginPageModule)
+    loadChildren: () => import('./view/login/login.module').then( m => m.LoginPageModule)
   },
   {
     path: 'signup',
-    loadChildren: () => import('./signup/signup.module').then( m => m.SignupPageModule)
+    loadChildren: () => import('./view/signup/signup.module').then( m => m.SignupPageModule)
   }
 ];
 @NgModule({
